Filtro para aqruivos .pdf na inserção de tutoriais do servidor + TryCatch para o id de lab em um chamado

+++ ProjectColab/ProjectColab/3-Bolsista/WebFormAddChamadoBolsista.aspx.cs
@@ -35,7 +35,7 @@ namespace ProjectColab._3_Bolsista
             Modelo.Chamado aChamado;
             DAL.DALChamado aDALChamado;
 
-            aChamado = new Modelo.Chamado("1", Session["idusuario"].ToString(), labDrop.Text, 1, "", Convert.ToString(resumo.Text), Convert.ToDecimal(quantidadeeq.Text), DateTime.Now);
+            aChamado = new Modelo.Chamado("1", Session["idusuario"].ToString(), labDrop.Text, 1, "", Convert.ToString(resumo.Text), Convert.ToInt32(quantidadeeq.Text), DateTime.Now);
             aDALChamado = new DAL.DALChamado();
 
             try
